-  handle exceptions in the order
